+++ lib/models/book_note.dart
@@ -7,6 +7,8 @@ class BookNote {
   String type;
   String color;
   String? readerNote;
+  String? contextPrefix;
+  String? contextSuffix;
   DateTime? createTime;
   DateTime updateTime;
 
@@ -23,6 +25,8 @@ class BookNote {
     required this.type,
     required this.color,
     this.readerNote,
+    this.contextPrefix,
+    this.contextSuffix,
     this.createTime,
     required this.updateTime,
   });
@@ -37,6 +41,8 @@ class BookNote {
       'type': type,
       'color': color,
       'reader_note': readerNote,
+      'context_prefix': contextPrefix,
+      'context_suffix': contextSuffix,
       'create_time': createTime?.toIso8601String(),
       'update_time': updateTime.toIso8601String(),
     };
@@ -49,6 +55,8 @@ class BookNote {
       'value': cfi,
       'type': type,
       'color': '#$color',
+      'contextPrefix': contextPrefix,
+      'contextSuffix': contextSuffix,
     };
   }
 
@@ -65,6 +73,8 @@ class BookNote {
       type: map['type'] as String? ?? '',
       color: map['color'] as String? ?? '',
       readerNote: map['reader_note'] as String?,
+      contextPrefix: map['context_prefix'] as String?,
+      contextSuffix: map['context_suffix'] as String?,
       createTime:
           createTimeString != null ? DateTime.tryParse(createTimeString) : null,
       updateTime: updateTimeString != null